add_geometry_support.py
"""Quick implementation plan for geometry support in HDI"""

print("""
🗺️ GEOMETRY SUPPORT IMPLEMENTATION PLAN
=====================================

QUICK ANSWER: This is NOT complex and won't slow us down!

1. BACKEND CHANGES (10 minutes):
   - Update PostgreSQL query to include geometry_wkt
   - Add geometry fields to response format
   - The data is already there!

2. DATA FLOW:
   Backend returns:
   {
     "property_address": "4118 Ella Blvd",
     "geometry": {
       "wkt": "POLYGON((-95.063383 29.699763...))",  
       "centroid": {"lat": 29.700916, "lon": -95.064231},
       "bbox": {
         "minX": -95.065078,
         "minY": 29.699727,
         "maxX": -95.063383, 
         "maxY": 29.702105
       }
     },
     ... other fields
   }

3. UI IMPLEMENTATION (Your choice):
   Option A: Leaflet.js (Simple, fast)
   Option B: Mapbox GL (More features)
   Option C: Google Maps (Familiar)
   
4. PERFORMANCE:
   - NO performance impact on backend
   - Geometry data loads with property (<10ms extra)
   - Map rendering is handled by browser

5. STREAMLIT EXAMPLE:
   ```python
   import folium
   import streamlit as st
   from shapely import wkt
   
   # Get property with geometry
   property_data = get_property("4118 Ella Blvd")
   
   # Create map
   m = folium.Map(location=[29.700916, -95.064231], zoom_start=18)
   
   # Add property polygon
   if property_data['geometry']['wkt']:
       shape = wkt.loads(property_data['geometry']['wkt'])
       folium.GeoJson(shape).add_to(m)
   
   # Display in Streamlit
   st_folium(m, width=700, height=500)
   ```

IMPLEMENTATION TIME: 30 minutes total
- Backend: 10 mins
- Basic UI: 20 mins

This will NOT slow down the project!
""")