requirements.txt
# Core Flask Dependencies
flask==3.0.3
flask-restx==1.3.0
flask-caching==2.3.0
flask-limiter==3.7.0
flask-cors==4.0.1
flask-compress==1.14
flasgger==0.9.7

# Perplexity Integration
openai==1.35.0

# Data Processing
pydantic==2.7.4
pandas==2.2.2
numpy==1.26.4

# Caching & Storage
redis==5.0.6

# Database
psycopg2-binary==2.9.9

# HTTP & Async
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.3

# Web Scraping (for HCAD)
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
selenium==4.15.0
webdriver-manager==4.0.1

# Environment & Config
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.2.0
prometheus-flask-exporter==0.23.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Production Server
gunicorn==22.0.0

# CLI Tools
tabulate==0.9.0

# Semantic Search (for future)
sentence-transformers==2.7.0
scikit-learn==1.5.0

# Task Queue (for future)
celery==5.3.4

# Development Tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# Streamlit Frontend
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0